requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
alembic==1.14.0
httpx==0.28.1
pydantic-settings==2.7.1
PyJWT==2.10.1
APScheduler==3.11.0
python-multipart==0.0.20
pymysql==1.1.1
pytest==8.3.4
pytest-asyncio==0.25.0

